requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.7.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.9.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.2.2
fasteners==0.14.1
frozenlist==1.4.1
future==0.18.2
Genshi==0.7.7
h11==0.14.0
httpcore==1.0.5
httplib2==0.14.0
httpx==0.27.0
idna==3.7
ippserver==0.2
joblib==1.4.0
keras==2.15.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libretranslatepy==2.1.1
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
multidict==6.0.5
mutagen==1.47.0
netifaces==0.10.4
nltk==3.8.1
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
openai==0.28.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
piston-mini-client==0.7.5
protobuf==3.6.1
pycairo==1.16.2
pycryptodomex==3.20.0
pycups==1.9.73
pydantic==2.8.2
pydantic-core==2.20.1
pydub==0.25.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
pytube==15.0.0
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.4.2
regex==2024.4.28
relatorio==0.10.2
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
soupsieve==2.5
system-service==0.3
systemd-python==234
textblob==0.18.0.post0
tqdm==4.66.2
translate==3.6.1
typing-extensions==4.11.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-devhelp==0.1
urllib3==2.2.1
usb-creator==0.3.7
wadllib==1.3.3
websockets==12.0
xdiagnose==3.8.9
xkit==0.0.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.4.9
